Rewrite `recheck` package in TypeScript
