+++ packages/recheck/package.json
@@ -16,18 +16,21 @@
     "vulnerability"
   ],
   "scripts": {
-    "build": "esbuild --bundle --platform=node --format=cjs src/index.js --outfile=main.js && esbuild --bundle --platform=browser --format=esm src/fallback.js --outfile=browser.js"
+    "build": "esbuild --bundle --platform=node --format=cjs src/index.ts --outfile=main.js && esbuild --bundle --platform=browser --format=esm src/fallback.ts --outfile=browser.js",
+    "typecheck": "tsc --noEmit -p ."
   },
   "files": [
     "browser.js",
     "main.js"
   ],
   "optionalDependencies": {
-    "recheck-macos-x64": "4.0.2",
     "recheck-linux-x64": "4.0.2",
+    "recheck-macos-x64": "4.0.2",
     "recheck-windows-x64": "4.0.2"
   },
   "devDependencies": {
-    "esbuild": "^0.13.15"
+    "@types/node": "^16.11.9",
+    "esbuild": "^0.13.15",
+    "typescript": "^4.5.2"
   }
 }
